commit changes

+++ src/Controller/AdminController.php
@@ -2,11 +2,14 @@
 
 namespace App\Controller;
 
+use App\Entity\Category;
 use App\Entity\Comment;
 use App\Entity\Project;
 use App\Entity\Idea;
 use App\Entity\User;
+use App\Form\CategoryType;
 use App\Form\UserType;
+use App\Repository\CategoryRepository;
 use App\Repository\ProjectRepository;
 use App\Repository\IdeaRepository;
 use App\Repository\CommentRepository;
@@ -28,14 +31,18 @@ class AdminController extends AbstractController
      * lists all projects in admin
      */
     #[Route('/project_list', name: 'app_admin_project_list', methods: ['GET'])]
-    #[IsGranted('ROLE_ADMIN')]
     public function projectList(
         ProjectRepository $projectRepository,
     ): Response {
-        $projects = $projectRepository->findAll();
-        return $this->render('admin/admin_project_list.html.twig', [
+        if ($this->isGranted('ROLE_ADMIN')) {
+            $projects = $projectRepository->findAll();
+            return $this->render('admin/admin_project_list.html.twig', [
             'projects' => $projects,
-        ]);
+            ]);
+        } else {
+            $this->addFlash('danger', 'You do not have access rights, please contact your administrator');
+            return $this->redirectToRoute('app_home', [], Response::HTTP_SEE_OTHER);
+        }
     }
 
     /**
@@ -44,14 +51,17 @@ class AdminController extends AbstractController
      * project view page for admin
      */
     #[Route('/project_view/{id}', name: 'app_admin_project_view', methods: ['GET'])]
-    #[IsGranted('ROLE_ADMIN')]
     public function projectView(
         Project $project,
     ): Response {
-
-        return $this->render('admin/admin_project_view.html.twig', [
-            'project' => $project,
-        ]);
+        if ($this->isGranted('ROLE_ADMIN')) {
+            return $this->render('admin/admin_project_view.html.twig', [
+                'project' => $project,
+            ]);
+        } else {
+            $this->addFlash('danger', 'You do not have access rights, please contact your administrator');
+            return $this->redirectToRoute('app_home', [], Response::HTTP_SEE_OTHER);
+        }
     }
 
     /**
@@ -62,14 +72,18 @@ class AdminController extends AbstractController
      * deletes project
      */
     #[Route('/delete_project/{id}', name: 'app_admin_project_delete', methods: ['POST'])]
-    #[IsGranted('ROLE_ADMIN')]
     public function projectDelete(Request $request, Project $project, ProjectRepository $projectRepository): Response
     {
-        if ($this->isCsrfTokenValid('delete' . $project->getId(), $request->request->get('_token'))) {
-            $projectRepository->remove($project, true);
+        if ($this->isGranted('ROLE_ADMIN')) {
+            if ($this->isCsrfTokenValid('delete' . $project->getId(), $request->request->get('_token'))) {
+                $projectRepository->remove($project, true);
+                $this->addFlash('notice', 'Project deleted');
+            }
+            return $this->redirectToRoute('app_admin_project_list', [], Response::HTTP_SEE_OTHER);
+        } else {
+            $this->addFlash('danger', 'You do not have access rights, please contact your administrator');
+            return $this->redirectToRoute('app_home', [], Response::HTTP_SEE_OTHER);
         }
-
-        return $this->redirectToRoute('app_admin_project_list', [], Response::HTTP_SEE_OTHER);
     }
 
     //    ---------------------------------------- Idea Routes -------------------------------------------------------
@@ -80,14 +94,18 @@ class AdminController extends AbstractController
      * lists all ideas in admin page
      */
     #[Route('/idea_list', name: 'app_admin_idea_list', methods: ['GET'])]
-    #[IsGranted('ROLE_ADMIN')]
     public function ideaList(
         IdeaRepository $ideaRepository,
     ): Response {
-        $ideas = $ideaRepository->findAll();
-        return $this->render('admin/admin_idea_list.html.twig', [
-            'ideas' => $ideas,
-        ]);
+        if ($this->isGranted('ROLE_ADMIN')) {
+            $ideas = $ideaRepository->findAll();
+            return $this->render('admin/admin_idea_list.html.twig', [
+                'ideas' => $ideas,
+            ]);
+        } else {
+            $this->addFlash('danger', 'You do not have access rights, please contact your administrator');
+            return $this->redirectToRoute('app_home', [], Response::HTTP_SEE_OTHER);
+        }
     }
 
     /**
@@ -96,13 +114,17 @@ class AdminController extends AbstractController
      * idea view page for admin
      */
     #[Route('/idea_view/{id}', name: 'app_admin_idea_view', methods: ['GET'])]
-    #[IsGranted('ROLE_ADMIN')]
     public function ideaView(
         Idea $idea,
     ): Response {
-        return $this->render('admin/admin_idea_view.html.twig', [
-            'idea' => $idea,
-        ]);
+        if ($this->isGranted('ROLE_ADMIN')) {
+            return $this->render('admin/admin_idea_view.html.twig', [
+                'idea' => $idea,
+            ]);
+        } else {
+            $this->addFlash('danger', 'You do not have access rights, please contact your administrator');
+            return $this->redirectToRoute('app_home', [], Response::HTTP_SEE_OTHER);
+        }
     }
 
     /**
@@ -113,14 +135,18 @@ class AdminController extends AbstractController
      * deletes idea
      */
     #[Route('/delete_idea/{id}', name: 'app_admin_idea_delete', methods: ['POST'])]
-    #[IsGranted('ROLE_ADMIN')]
     public function ideaDelete(Request $request, Idea $idea, IdeaRepository $ideaRepository): Response
     {
-        if ($this->isCsrfTokenValid('delete' . $idea->getId(), $request->request->get('_token'))) {
-            $ideaRepository->remove($idea, true);
+        if ($this->isGranted('ROLE_ADMIN')) {
+            if ($this->isCsrfTokenValid('delete' . $idea->getId(), $request->request->get('_token'))) {
+                $ideaRepository->remove($idea, true);
+                $this->addFlash('notice', 'Idea deleted');
+            }
+            return $this->redirectToRoute('app_admin_idea_list', [], Response::HTTP_SEE_OTHER);
+        } else {
+            $this->addFlash('danger', 'You do not have access rights, please contact your administrator');
+            return $this->redirectToRoute('app_home', [], Response::HTTP_SEE_OTHER);
         }
-
-        return $this->redirectToRoute('app_admin_idea_list', [], Response::HTTP_SEE_OTHER);
     }
 
     //---------------------------------------- Comment Routes -------------------------------------------------------
@@ -131,14 +157,18 @@ class AdminController extends AbstractController
      * lists all comments
      */
     #[Route('/comment_list', name: 'app_admin_comment_list', methods: ['GET'])]
-    #[IsGranted('ROLE_ADMIN')]
     public function commentList(
         CommentRepository $commentRepository,
     ): Response {
-        $comments = $commentRepository->findAll();
-        return $this->render('admin/admin_comment_list.html.twig', [
-            'comments' => $comments,
-        ]);
+        if ($this->isGranted('ROLE_ADMIN')) {
+            $comments = $commentRepository->findAll();
+            return $this->render('admin/admin_comment_list.html.twig', [
+                'comments' => $comments,
+            ]);
+        } else {
+            $this->addFlash('danger', 'You do not have access rights, please contact your administrator');
+            return $this->redirectToRoute('app_home', [], Response::HTTP_SEE_OTHER);
+        }
     }
 
     /**
@@ -148,14 +178,18 @@ class AdminController extends AbstractController
      * comment view page for admin
      */
     #[Route('/comment_view/{id}', name: 'app_admin_comment_view', methods: ['GET'])]
-    #[IsGranted('ROLE_ADMIN')]
     public function commentView(
         Comment $comment,
         IdeaRepository $ideaRepository,
     ): Response {
-        return $this->render('admin/admin_comment_view.html.twig', [
-            'comment' => $comment,
-        ]);
+        if ($this->isGranted('ROLE_ADMIN')) {
+            return $this->render('admin/admin_comment_view.html.twig', [
+                'comment' => $comment,
+            ]);
+        } else {
+            $this->addFlash('danger', 'You do not have access rights, please contact your administrator');
+            return $this->redirectToRoute('app_home', [], Response::HTTP_SEE_OTHER);
+        }
     }
 
     /**
@@ -166,13 +200,18 @@ class AdminController extends AbstractController
      * deletes comment
      */
     #[Route('/delete_comment/{id}', name: 'app_admin_comment_delete', methods: ['POST'])]
-    #[IsGranted('ROLE_ADMIN')]
     public function deleteComment(Request $request, Comment $comment, CommentRepository $commentRepository): Response
     {
-        if ($this->isCsrfTokenValid('delete' . $comment->getId(), $request->request->get('_token'))) {
-            $commentRepository->remove($comment, true);
+        if ($this->isGranted('ROLE_ADMIN')) {
+            if ($this->isCsrfTokenValid('delete' . $comment->getId(), $request->request->get('_token'))) {
+                $commentRepository->remove($comment, true);
+                $this->addFlash('notice', 'Comment deleted');
+            }
+            return $this->redirectToRoute('app_admin_comment_list', [], Response::HTTP_SEE_OTHER);
+        } else {
+            $this->addFlash('danger', 'You do not have access rights, please contact your administrator');
+            return $this->redirectToRoute('app_home', [], Response::HTTP_SEE_OTHER);
         }
-        return $this->redirectToRoute('app_admin_comment_list', [], Response::HTTP_SEE_OTHER);
     }
 
     // ---------------------------------------- User Routes -------------------------------------------------------
@@ -182,14 +221,17 @@ class AdminController extends AbstractController
      * @return Response
      * lists all users
      */
-    #[Route('/', name: 'app_admin_user_list', methods: ['GET'])]
-    #[IsGranted('ROLE_ADMIN')]
+    #[Route('/users', name: 'app_admin_user_list', methods: ['GET'])]
     public function index(UserRepository $userRepository): Response
     {
-
-        return $this->render('admin/admin_user_list.html.twig', [
-            'users' => $userRepository->findAll(),
-        ]);
+        if ($this->isGranted('ROLE_ADMIN')) {
+            return $this->render('admin/admin_user_list.html.twig', [
+                'users' => $userRepository->findAll(),
+            ]);
+        } else {
+            $this->addFlash('danger', 'You do not have access rights, please contact your administrator');
+            return $this->redirectToRoute('app_home', [], Response::HTTP_SEE_OTHER);
+        }
     }
 
     /**
@@ -199,24 +241,28 @@ class AdminController extends AbstractController
      * adds new user
      */
     #[Route('/new_user', name: 'app_admin_user_new', methods: ['GET', 'POST'])]
-    #[IsGranted('ROLE_ADMIN')]
     public function new(Request $request, UserRepository $userRepository): Response
     {
-        $user = new User();
-        $form = $this->createForm(UserType::class, $user);
-        $form->handleRequest($request);
-        if ($form->isSubmitted() && $form->isValid()) {
-            $passwordUser = '1234';
-            $user->setPassword(password_hash($passwordUser, PASSWORD_DEFAULT));
-            $userRepository->save($user, true);
-            return $this->redirectToRoute('app_admin_user_list', [], Response::HTTP_SEE_OTHER);
+        if ($this->isGranted('ROLE_ADMIN')) {
+            $user = new User();
+            $form = $this->createForm(UserType::class, $user);
+            $form->handleRequest($request);
+            if ($form->isSubmitted() && $form->isValid()) {
+                $passwordUser = '1234';
+                $user->setPassword(password_hash($passwordUser, PASSWORD_DEFAULT));
+                $userRepository->save($user, true);
+                $this->addFlash('success', 'Success: User added - email sent');
+                return $this->redirectToRoute('app_admin_user_list', [], Response::HTTP_SEE_OTHER);
+            }
+            return $this->renderForm('admin/admin_user_new.html.twig', [
+                'user' => $user,
+                'form' => $form,
+                'edit' => true,
+            ]);
+        } else {
+            $this->addFlash('danger', 'You do not have access rights, please contact your administrator');
+            return $this->redirectToRoute('app_home', [], Response::HTTP_SEE_OTHER);
         }
-
-        return $this->renderForm('admin/admin_user_new.html.twig', [
-            'user' => $user,
-            'form' => $form,
-            'edit' => true,
-        ]);
     }
 
     /**
@@ -226,14 +272,104 @@ class AdminController extends AbstractController
      * @return Response
      * deletes user
      */
-    #[Route('/{id}', name: 'app_admin_user_delete', methods: ['POST'])]
-    #[IsGranted('ROLE_ADMIN')]
+    #[Route('/delete_user/{id}', name: 'app_admin_user_delete', methods: ['POST'])]
     public function deleteUser(Request $request, User $user, UserRepository $userRepository): Response
     {
-        if ($this->isCsrfTokenValid('delete' . $user->getId(), $request->request->get('_token'))) {
-            $userRepository->remove($user, true);
+        if ($this->isGranted('ROLE_ADMIN')) {
+            if ($this->isCsrfTokenValid('delete' . $user->getId(), $request->request->get('_token'))) {
+                $userRepository->remove($user, true);
+                $this->addFlash('notice', 'User - ' . $user->getFullName() . ' deleted');
+            }
+            return $this->redirectToRoute('app_admin_user_list', [], Response::HTTP_SEE_OTHER);
+        } else {
+            $this->addFlash('danger', 'You do not have access rights, please contact your administrator');
+            return $this->redirectToRoute('app_home', [], Response::HTTP_SEE_OTHER);
         }
+    }
 
-        return $this->redirectToRoute('app_admin_user_list', [], Response::HTTP_SEE_OTHER);
+    // ---------------------------------------- Category Routes -------------------------------------------------------
+
+    /**
+     * @param Request $request
+     * @param CategoryRepository $categoryRepository
+     * @return Response
+     * department admin page, renders form on same page to add departments
+     */
+    #[Route('/departments', name: 'app_admin_category_index', methods: ['GET', 'POST'])]
+    public function indexCategory(Request $request, CategoryRepository $categoryRepository): Response
+    {
+        if ($this->isGranted('ROLE_ADMIN')) {
+            $category = new Category();
+            $form = $this->createForm(CategoryType::class, $category);
+            $form->handleRequest($request);
+            if ($form->isSubmitted() && $form->isValid()) {
+                $categoryRepository->save($category, true);
+                $this->addFlash('success', 'Success: Department created');
+                return $this->redirectToRoute('app_admin_category_index', [], Response::HTTP_SEE_OTHER);
+            }
+            return $this->renderform('admin/admin_category_list.html.twig', [
+                'categories' => $categoryRepository->findAll(),
+                'form' => $form,
+                'edit' => true,
+            ]);
+        } else {
+            $this->addFlash('danger', 'You do not have access rights, please contact your administrator');
+            return $this->redirectToRoute('app_home', [], Response::HTTP_SEE_OTHER);
+        }
+    }
+
+    /**
+     * @param Request $request
+     * @param Category $category
+     * @param CategoryRepository $categoryRepository
+     * @return Response
+     * lists all departments and renders form on same page to edit department
+     */
+    #[Route('/edit_department/{id}', name: 'app_admin_category_edit', methods: ['GET', 'POST'])]
+    public function editCategory(Request $request, Category $category, CategoryRepository $categoryRepository): Response
+    {
+        if ($this->isGranted('ROLE_ADMIN')) {
+            $form = $this->createForm(CategoryType::class, $category);
+            $form->handleRequest($request);
+            if ($form->isSubmitted() && $form->isValid()) {
+                $categoryRepository->save($category, true);
+                $this->addFlash('success', "Success: Category " . "'" . $category->getTitle() . "' modified");
+                return $this->redirectToRoute('app_admin_category_index', [], Response::HTTP_SEE_OTHER);
+            }
+            return $this->renderForm('admin/admin_category_edit.html.twig', [
+                'categories' => $categoryRepository->findAll(),
+                'categoryById' => $categoryRepository->findOneBy(['id' => $category->getId()]),
+                'form' => $form,
+                'edit' => true,
+            ]);
+        } else {
+            $this->addFlash('danger', 'You do not have access rights, please contact your administrator');
+            return $this->redirectToRoute('app_home', [], Response::HTTP_SEE_OTHER);
+        }
+    }
+
+    /**
+     * @param Request $request
+     * @param Category $category
+     * @param CategoryRepository $categoryRepository
+     * @return Response
+     * deletes department
+     */
+    #[Route('/delete_department/{id}', name: 'app_admin_category_delete', methods: ['POST'])]
+    public function deleteCategory(
+        Request $request,
+        Category $category,
+        CategoryRepository $categoryRepository
+    ): Response {
+        if ($this->isGranted('ROLE_ADMIN')) {
+            if ($this->isCsrfTokenValid('delete' . $category->getId(), $request->request->get('_token'))) {
+                $categoryRepository->remove($category, true);
+                $this->addFlash('notice', "Department '" . $category->getTitle() . ' deleted');
+            }
+            return $this->redirectToRoute('app_admin_category_index', [], Response::HTTP_SEE_OTHER);
+        } else {
+            $this->addFlash('danger', 'You do not have access rights, please contact your administrator');
+            return $this->redirectToRoute('app_home', [], Response::HTTP_SEE_OTHER);
+        }
     }
 }
